Removed temporary import channel-bumps command
The channel bumps will now automatically be imported whenever there is a
report that contains a yet unknown bump.

As part of that I also changed the database field of the advance date to
be an Option<String> as that is what the reports currently look like. At
some point I might want to regenerate all the old ones so I can get rid
of the snowflakes.