it is time to support proper version matching
Looking throught the current JSON files there seems to be only LT and EQ
in the `version_affected` fields. I implemented both and also
implemented version splitting in the sense of how Nix does it. I hope
that is good enough for now. I'd love to just be able to compare strings
but (especially) the `pre` case in the string is explicitly handled
differently…